# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitCheck engine",
    "summary": "Scan a garment, see it on you, get BUY, SKIP or TRY-WITH.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/scan": {
      "post": {
        "summary": "Scan",
        "description": "Cut out and tag one garment photo. Nothing is stored.",
        "operationId": "scan_scan_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_scan_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/link": {
      "post": {
        "summary": "Link",
        "description": "Fetch the garment image from a shop or image link, to scan a garment not owned yet.",
        "operationId": "link_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/judge": {
      "post": {
        "summary": "Judge",
        "description": "Return BUY, SKIP or TRY_WITH for scanned tags against the owner's closet and week.",
        "operationId": "judge_judge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JudgeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/week/{owner}": {
      "get": {
        "summary": "Week",
        "description": "Return the forecast and calendar events for the coming days.",
        "operationId": "week_week__owner__get",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/render": {
      "post": {
        "summary": "Render",
        "description": "Render the garment on the person.",
        "operationId": "render_render_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_render_render_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "summary": "Chat",
        "description": "Ask the stylist about the current scan, verdict or closet.",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/closet/{owner}": {
      "get": {
        "summary": "Closet",
        "description": "List the owner's closet.",
        "operationId": "closet_closet__owner__get",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Garment-Output"
                  },
                  "title": "Response Closet Closet  Owner  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add To Closet",
        "description": "Add a garment photo to the closet, tagging it unless `tags_json` is given.",
        "operationId": "add_to_closet_closet__owner__post",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_to_closet_closet__owner__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Forget",
        "description": "Delete every garment and stored image for the owner.",
        "operationId": "forget_closet__owner__delete",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/closet/{owner}/scan": {
      "post": {
        "summary": "Scan Closet",
        "description": "Find every garment in a photo of a rack or closet and add them all.",
        "operationId": "scan_closet_closet__owner__scan_post",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_closet_closet__owner__scan_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetScanOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/closet/{owner}/{garment_id}/image": {
      "get": {
        "summary": "Garment Image",
        "description": "Return the stored garment cutout as PNG.",
        "operationId": "garment_image_closet__owner___garment_id__image_get",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner"
            }
          },
          {
            "name": "garment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Garment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddOut": {
        "properties": {
          "garment": {
            "$ref": "#/components/schemas/Garment-Output"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "garment",
          "pipeline"
        ],
        "title": "AddOut"
      },
      "Body_add_to_closet_closet__owner__post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          },
          "tags_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json",
            "description": "GarmentTags as JSON; omit to auto-tag"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "source": {
            "$ref": "#/components/schemas/Source",
            "default": "closet"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "Shop link it came from"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_add_to_closet_closet__owner__post"
      },
      "Body_render_render_post": {
        "properties": {
          "person": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Person",
            "description": "Person photo; held in memory only"
          },
          "garment": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Garment",
            "description": "Garment photo or cutout"
          },
          "region": {
            "$ref": "#/components/schemas/TryOnRegion"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "required": [
          "person",
          "garment",
          "region"
        ],
        "title": "Body_render_render_post"
      },
      "Body_scan_closet_closet__owner__scan_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_scan_closet_closet__owner__scan_post"
      },
      "Body_scan_scan_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_scan_scan_post"
      },
...
      "ChatIn": {
        "properties": {
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatTurn"
            },
            "type": "array",
            "title": "History",
            "default": []
          },
          "candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GarmentTags"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "owner",
          "message"
        ],
        "title": "ChatIn"
      },
      "ChatOut": {
        "properties": {
          "reply": {
            "$ref": "#/components/schemas/StylistReply"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "reply",
          "pipeline"
        ],
        "title": "ChatOut"
      },
...
      "ClosetScanOut": {
        "properties": {
          "garments": {
            "items": {
              "$ref": "#/components/schemas/Garment-Output"
            },
            "type": "array",
            "title": "Garments"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "garments",
          "pipeline"
        ],
        "title": "ClosetScanOut"
      },
...
      "ForgetOut": {
        "properties": {
          "deleted": {
            "type": "integer",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "ForgetOut"
      },
...
      "Garment-Output": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "source": {
            "$ref": "#/components/schemas/Source",
            "default": "closet"
          },
          "tags": {
            "$ref": "#/components/schemas/GarmentTags"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "wears": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Wears",
            "default": 0
          },
          "image_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Ref"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "owner",
          "tags"
        ],
        "title": "Garment",
        "description": "A tagged garment that belongs to one owner's closet."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JudgeIn": {
        "properties": {
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "tags": {
            "$ref": "#/components/schemas/GarmentTags"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "owner",
          "tags"
        ],
        "title": "JudgeIn"
      },
      "JudgeOut": {
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/Verdict-Output"
          },
          "week": {
            "$ref": "#/components/schemas/WeekContext"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "week",
          "pipeline"
        ],
        "title": "JudgeOut"
      },
      "LinkIn": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "LinkIn"
      },
      "LinkOut": {
        "properties": {
          "image_png_base64": {
            "type": "string",
            "title": "Image Png Base64"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "image_png_base64",
          "source_url",
          "title",
          "pipeline"
        ],
        "title": "LinkOut"
      },
...
      "RenderOut": {
        "properties": {
          "image_png_base64": {
            "type": "string",
            "title": "Image Png Base64"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "image_png_base64",
          "cached",
          "pipeline"
        ],
        "title": "RenderOut"
      },
...
      "ScanOut": {
        "properties": {
          "tags": {
            "$ref": "#/components/schemas/GarmentTags"
          },
          "cutout_png_base64": {
            "type": "string",
            "title": "Cutout Png Base64"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "tags",
          "cutout_png_base64",
          "pipeline"
        ],
        "title": "ScanOut"
      },
...
      "WeekOut": {
        "properties": {
          "week": {
            "$ref": "#/components/schemas/WeekContext"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            },
            "type": "array",
            "title": "Pipeline"
          }
        },
        "type": "object",
        "required": [
          "week",
          "pipeline"
        ],
        "title": "WeekOut"
      }
    }
  }
}